+++ packages/mochi-formatter/src/string.ts
@@ -1,5 +1,7 @@
 import { HOMEPAGE } from "./constant";
 import address from "./address";
+import type { AssociatedAccount } from "@consolelabs/mochi-rest";
+import { Platform } from "./ui";
 
 function receiptLink(id: string, wrapCode = false) {
   return `[${wrapCode ? "`" : ""}${id.slice(0, 5)}${
@@ -7,15 +9,40 @@ function receiptLink(id: string, wrapCode = false) {
   }](${HOMEPAGE}/tx/${id})`;
 }
 
-function formatAddressUsername(value: string, len = 15) {
+function formatAddressUsername(value: string | AssociatedAccount, len = 15) {
   const _len = Math.min(Math.max(len, 0), 20);
-  const isAddressResult = address.isAddress(value);
-  if (isAddressResult.valid && !address.isShorten(value)) {
+
+  if (typeof value !== "string") {
+    return _formatDomainName(value);
+  }
+
+  if (address.isAddress(value).valid && !address.isShorten(value)) {
     return address.shorten(value);
   }
-  const hidden = value.slice(_len);
 
-  return `${value.slice(0, _len)}${hidden.length ? "..." : ""}`;
+  return _formatUsername(value, _len);
+}
+
+function _formatDomainName(p: AssociatedAccount) {
+  const isSolAcc = p.platform === Platform.SolanaChain;
+
+  if (isSolAcc) {
+    return p.platform_metadata.sns || address.shorten(p.platform_identifier);
+  }
+
+  const isEvmAcc = p.platform === Platform.EvmChain;
+
+  if (isEvmAcc) {
+    return p.platform_metadata.ens || address.shorten(p.platform_identifier);
+  }
+
+  return address.shorten(p.platform_identifier);
+}
+
+function _formatUsername(value: string, len: number) {
+  const hidden = value.slice(len);
+
+  return `${value.slice(0, len)}${hidden.length ? "..." : ""}`;
 }
 
 export default {